pyproject.toml
[project]
name = "langgraph-multiagent"
version = "0.1.0"
description = "Multi-agent system with LangGraph: tool routing and conversational memory"
authors = [{name = "Salma Ouardi", email = "salmaouardi1998@gmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",
    "ollama>=0.1.7",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "typing_extensions>=4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
ui = [
    "streamlit>=1.29.0",
]
api = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
